Fix Stable LoRA Weighting
This fix ensures that you will get an accurate representation of the data you are training when saving your LoRA file. 

There are cases where your data may overfit during training, but will not show until the weights are fully merged, which is presumed to most likely happen when you are ready for inference (webui, inference script, etc). `merged` and `merge_weights` represent two different types of functionality.

`merge_weights`:  Allow the model to merge or unmerge the current LoRA weights that are set. If this parameter is set to False, the LoRA weights will still be processed, but won't show an accurate representation of how it would look at inference time.

`merged`: The weights are merged, run as if we are in inference. Not merged, this is assumed to be training.