Prevent Conv3d LoRA Merging
It's possible that on validation, adding the mean then removing it could cause a vanishing gradient problem where the backpropogation could return NaN values due to the Conv3d LoRA layer. 

The fix is to only allow forward passes for Conv3d LoRA during training, but allow the user to merge themselves by manually setting the flag after training. Still in testing, but most likely to be added as a stability enhancement.